fix(prysm): run p2p-udp on different port as it might conflict with the new quic port (#845)
Prysm added QUIC to their client as per
https://github.com/prysmaticlabs/prysm/pull/14688 . This currently broke
our runs using the master branch prysm image due to the QUIC port
running on 13000 by default. Which was conflicting with the settings we
were providing for the UDP port:

```sh
--p2p-quic-port value                                The QUIC port used by libp2p. (default: 13000)
--p2p-tcp-port value                                 The TCP port used by libp2p. (default: 13000)
--p2p-udp-port value                                 The UDP port used by the discovery service discv5. (default: 12000)
```

I've also added the CLI flags for it, but still commented out until a
stable release is done.