feat: add Tempo as an additional service to collect Lighthouse tracing data (#1150)
This PR adds `tempo` as an additional service, and adds the collector
url to lighthouse startup parameters, so that lighthouse nodes can push
traces to Tempo tracing backend when it's enabled.

When tempo is enabled, Lighthouse traces can be observed in the Grafana
`Explore` tab, and select `Search` query type:

<img width="1071" height="765" alt="image"
src="https://github.com/user-attachments/assets/bbcd7704-17e9-4d0e-bce4-69da6f07b9f7"
/>

I have a separate branch identical to this one but with the
`kurtosis.yml` file modified to my fork, so it can be run with:
```
kurtosis run --enclave local-testnet github.com/jimmygchen/ethereum-package@tempo-test --args-file ./network_params.yaml
```


Note: this PR is coded by Claude and it's working - I'm using it but I
likely won't have capacity in the short term to polish it further, feel
free to leave the PR open or take over.

Another current limitation is that the service name is hard coded to
`lighthouse` so we can't distinguish between different lighthouse nodes.
We could add a flag to lighthouse to configure the tracing service name.